requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

python-etcd
python-dateutil
gevent>=1.0
greenlet>=0.3.2
msgpack-python
pytz
